Tweaks. Fixes. Additional tests.

+++ test/dev/kkorolyov/ezdb/connection/concrete/SimpleDatabaseConnectionTest.java
@@ -7,27 +7,44 @@ import java.sql.SQLException;
 
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
 import dev.kkorolyov.ezdb.connection.DatabaseConnection;
 import dev.kkorolyov.ezdb.construct.Column;
+import dev.kkorolyov.ezdb.construct.RowEntry;
 import dev.kkorolyov.ezdb.construct.SqlType;
 import dev.kkorolyov.ezdb.exceptions.ClosedException;
 import dev.kkorolyov.ezdb.exceptions.DuplicateTableException;
+import dev.kkorolyov.ezdb.exceptions.MismatchedTypeException;
 import dev.kkorolyov.ezdb.exceptions.NullTableException;
+import dev.kkorolyov.ezdb.logging.DebugLogger;
+import dev.kkorolyov.ezdb.properties.Properties;
 
 @SuppressWarnings("javadoc")
 public class SimpleDatabaseConnectionTest {
-	private static final String TEST_HOST = "192.168.1.157", TEST_DB = "TEST_DB", TEST_USER = "postgres", TEST_PASSWORD = "", TEST_TABLE = "TEST_TABLE";
+	private static final String host = Properties.getValue(Properties.HOST), database = "TEST_DB", user = Properties.getValue(Properties.USER), password = Properties.getValue(Properties.PASSWORD),table = "DATABASE_TEST_TABLE";
 	
 	private DatabaseConnection conn;
+	private Column[] columns = {new Column("TEST_COL_1", SqlType.BOOLEAN)};
 	
+	@BeforeClass
+	public static void setUpBeforeClass() {
+		DebugLogger.enableAll();
+	}
 	@Before
-	public void setUp() throws SQLException, DuplicateTableException, NullTableException {		
-		conn = new SimpleDatabaseConnection(TEST_HOST, TEST_DB, TEST_USER, TEST_PASSWORD);	// Use a fresh connection for each test
+	public void setUp() throws SQLException, DuplicateTableException, NullTableException, ClosedException {
+		conn = new SimpleDatabaseConnection(host, database, user, password);	// Use a fresh connection for each test
+		
+		if (conn.containsTable(table))
+			conn.dropTable(table);
+		conn.createTable(table, columns);
 	}
 	@After
-	public void tearDown() throws NullTableException, SQLException {	
+	public void tearDown() throws NullTableException, SQLException, ClosedException {
+		if (conn.isClosed())
+			conn = new SimpleDatabaseConnection(host, database, user, password);
+		conn.dropTable(table);
 		conn.close();	// Make sure all resources release after each test
 	}
 	
@@ -50,14 +67,14 @@ public class SimpleDatabaseConnectionTest {
 	}
 
 	@Test
-	public void testExecute() throws SQLException, ClosedException {
-		String testStatement = "SELECT * FROM " + TEST_TABLE;
+	public void testExecute() throws SQLException, ClosedException {	// Mainly for exceptions
+		String testStatement = "SELECT * FROM " + table;
 		conn.execute(testStatement);
 	}
 	@Test
-	public void testExecuteParams() {
-		String testStatement = "SELECT"
-		fail("Unimplemented");
+	public void testExecuteParams() throws SQLException, ClosedException, MismatchedTypeException {	// Mainly for exceptions
+		String testStatement = "SELECT " + columns[0].getName() + " FROM " + table + " WHERE " + columns[0].getName() + "=?";
+		conn.execute(testStatement, new RowEntry[]{new RowEntry(columns[0], false)});
 	}
 	
 	@Test
@@ -67,11 +84,8 @@ public class SimpleDatabaseConnectionTest {
 		if (conn.containsTable(testTableName))	// Clear stale test table from a previous run, if exists
 			conn.dropTable(testTableName);
 		
-		SqlType[] types = SqlType.values();
-		Column[] columns = new Column[types.length];	// Test all column types
-		for (int i = 0; i < columns.length; i++) {
-			columns[i] = new Column("TEST_COLUMN_" + types[i].getTypeName(), types[i]);
-		}
+		Column[] columns = buildAllColumns();
+		
 		assertTrue(!conn.containsTable(testTableName));
 		conn.createTable(testTableName, columns);
 		assertTrue(conn.containsTable(testTableName));
@@ -96,7 +110,8 @@ public class SimpleDatabaseConnectionTest {
 	public void testContainsTable() throws DuplicateTableException, SQLException, NullTableException, ClosedException {
 		String testTableName = "TEST_TABLE_CONTAINS";
 		
-		conn.dropTable(testTableName);
+		if (conn.containsTable(testTableName))
+			conn.dropTable(testTableName);
 		
 		assertTrue(!conn.containsTable(testTableName));
 		conn.createTable(testTableName, new Column[]{new Column("TEST_COLUMN", SqlType.BOOLEAN)});
@@ -110,5 +125,15 @@ public class SimpleDatabaseConnectionTest {
 		for (String table : conn.getTables()) {
 			System.out.println(table);
 		}
+	}
+	
+	private static Column[] buildAllColumns() {
+		SqlType[] allTypes = SqlType.values();
+		Column[] allColumns = new Column[allTypes.length];
+		
+		for (int i = 0; i < allColumns.length; i++)
+			allColumns[i] = new Column(allTypes[i].toString(), allTypes[i]);
+		
+		return allColumns;
 	}
 }
